Prepare first teest for apt list
