+++ handlers_containers_functional_test.go
@@ -23,20 +23,6 @@ import (
 	"golang.org/x/net/context"
 )
 
-// NewMqttTestClientLocal creates mqtt client in localhost:1883
-func NewMqttTestClientLocal() *MqttTestClient {
-	uiOptions := mqtt.NewClientOptions().AddBroker("tcp://localhost:1883").SetClientID("UI")
-	ui := mqtt.NewClient(uiOptions)
-	if token := ui.Connect(); token.Wait() && token.Error() != nil {
-		panic(token.Error())
-	}
-
-	return &MqttTestClient{
-		ui,
-		"",
-	}
-}
-
 type testStatus struct {
 	appStatus *Status
 	ui        *MqttTestClient
