+++ tests_helper.go
@@ -68,6 +68,20 @@ func NewMqttTestClient() *MqttTestClient {
 	}
 }
 
+// NewMqttTestClientLocal creates mqtt client in localhost:1883
+func NewMqttTestClientLocal() *MqttTestClient {
+	uiOptions := mqtt.NewClientOptions().AddBroker("tcp://localhost:1883").SetClientID("UI")
+	ui := mqtt.NewClient(uiOptions)
+	if token := ui.Connect(); token.Wait() && token.Error() != nil {
+		panic(token.Error())
+	}
+
+	return &MqttTestClient{
+		ui,
+		"",
+	}
+}
+
 // Close disconnect client
 func (tmc *MqttTestClient) Close() {
 	tmc.client.Disconnect(100)
